refactor App directory
