+++ pwa/src/containers/App/App.js
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import logo from "./logo.svg";
+import logo from "../../assets/img/logo.svg";
 import "./App.scss";
 
 class App extends Component {
